requirements.txt
email-validator>=2.3.0
flask-login>=0.6.3
flask>=3.1.2
flask-sqlalchemy>=3.1.1
flask-socketio>=5.3.6
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.43
werkzeug>=3.1.3
oauthlib>=3.3.1
pyjwt>=2.10.1
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
flask-login>=0.6.3
reportlab>=4.4.3
requests>=2.31.0
yfinance>=0.2.28
beautifulsoup4>=4.12.2
pandas>=2.1.0
python-dotenv>=1.0.0
lxml>=4.9.3
